pyproject.toml
[tool.poetry]
name = "codinglm"
version = "0.1.0"
description = "CodinGLM – a Claude Code clone powered by Z.ai's GLM models"
authors = ["Dustin Painter"]
readme = "README.md"
packages = [{include = "codinglm"}]

[tool.poetry.dependencies]
python = "^3.10"
zhipuai = "^2.0.0"  # Official Z.ai SDK
rich = "^13.7.0"  # Terminal UI with markdown rendering
click = "^8.1.7"  # CLI framework
httpx = "^0.27.0"  # HTTP client for web tools
pydantic = "^2.6.0"  # Data validation
python-dotenv = "^1.0.0"  # Environment variables
aiofiles = "^23.2.1"  # Async file operations
prompt-toolkit = "^3.0.43"  # Enhanced REPL
pygments = "^2.17.0"  # Syntax highlighting
gitpython = "^3.1.41"  # Git operations
beautifulsoup4 = "^4.12.3"  # HTML parsing for WebFetch
markdownify = "^0.12.1"  # HTML to Markdown
mcp = "^1.0.0"  # Model Context Protocol
tiktoken = {version = "^0.8.0", optional = true}  # Accurate token counting

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.3"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
codinglm = "codinglm.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
